requirements.txt
fastapi
uvicorn
pillow
rembg
python-multipart